Revert extracting target find logic
This reverts commit 036077076fd4a70010255a90008f499ba222a011.
Revert "expose renderTarget"
e5871f821
This reverts commit bca235587e57ef7aaea4d9b6b8bc29f37d436d01.
