+++ test/unit/canvas_dispose.js
@@ -316,9 +316,9 @@ function testCanvasDisposing() {
         assert.equal(canvas.wrapperEl, null, 'wrapperEl should be deleted');
         assert.equal(canvas.upperCanvasEl, null, 'upperCanvas should be deleted');
         assert.equal(canvas.lowerCanvasEl, null, 'lowerCanvasEl should be deleted');
-        assert.equal(canvas.targetFind.el, null, 'pixelFindCanvasEl should be deleted');
+        assert.equal(canvas.pixelFindCanvasEl, null, 'pixelFindCanvasEl should be deleted');
         assert.equal(canvas.contextTop, null, 'contextTop should be deleted');
-        assert.equal(canvas.targetFind.ctx, null, 'pixelFindContext should be deleted');
+        assert.equal(canvas.pixelFindContext, null, 'pixelFindContext should be deleted');
         assert.equal(canvas.elements._originalCanvasStyle, undefined, 'removed original canvas style');
         assert.equal(el.style.cssText, elStyle, 'restored original canvas style');
         assert.equal(el.width, 200, 'restored width');
